+++ storybook/.babelrc
@@ -1,0 +1,3 @@
+{
+    "presets": ["babel-preset-expo"]
+}
