+++ storybook/app.json
@@ -1,0 +1,29 @@
+{
+  "expo": {
+    "name": "React Native Responsive Layout Example",
+    "description":
+      "Example app for React Native Responsive Layout: https://github.com/axilis/react-native-responsive-layout",
+    "slug": "react-native-responsive-layout-example",
+    "privacy": "public",
+    "sdkVersion": "32.0.0",
+    "platforms": [
+      "ios",
+      "android"
+    ],
+    "version": "1.0.0",
+    "orientation": "default",
+    "splash": {
+      "resizeMode": "contain",
+      "backgroundColor": "#ffffff"
+    },
+    "updates": {
+      "fallbackToCacheTimeout": 0
+    },
+    "assetBundlePatterns": [
+      "**/*"
+    ],
+    "ios": {
+      "supportsTablet": true
+    }
+  }
+}
