+++ storybook/package.json
@@ -1,0 +1,21 @@
+{
+  "main": "node_modules/expo/AppEntry.js",
+  "scripts": {
+    "start": "expo start"
+  },
+  "dependencies": {
+    "@storybook/react-native": "^4.1.11",
+    "escape-string-regexp": "^1.0.5",
+    "expo": "^32.0.0",
+    "metro-config": "^0.51.1",
+    "react": "16.5.0",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz"
+  },
+  "devDependencies": {
+    "babel-preset-expo": "^5.0.0"
+  },
+  "resolutions": {
+    "react": "16.5.0"
+  },
+  "private": true
+}
